# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SME Finance Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{profileId}/income-tax-estimate:
    get:
      tags: [Tax]
      operationId: getIncomeTaxEstimate
      summary: Get the live, current-tax-year estimated income-tax position
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      responses:
        '200':
          description: Tax estimate based on canonical Financial Memory records and saved profile inputs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomeTaxEstimateResponse'
        '404':
          description: Profile not found or not owned by the current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /profiles/{profileId}/inbox:
    get:
      tags: [Inbox]
      operationId: listInboxItems
      summary: List inbox items for a profile
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      responses:
        '200':
          description: List of inbox items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InboxItem'
  /profiles/{profileId}/inbox/{itemId}/resolve:
    patch:
      tags: [Inbox]
      operationId: resolveInboxItem
      summary: Resolve an inbox item — creates a transaction and updates financial position
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InboxResolveInput'
      responses:
        '200':
          description: Inbox item resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxItem'
...
  /copilot/message:
    post:
      tags: [Copilot]
      operationId: sendCopilotMessage
      summary: Send a message to the AI copilot with live financial context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopilotMessageInput'
      responses:
        '200':
          description: AI response with financial context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CopilotMessageResponse'
...
components:
  parameters:
...
    ProfileIdParam:
      name: profileId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    IncomeTaxEstimateResponse:
      type: object
      required: [period, taxYear, accountingBasis, profitLoss, categories, estimate]
      properties:
        period:
          $ref: '#/components/schemas/TaxPeriod'
        taxYear:
          type: string
        accountingBasis:
          type: string
          enum: [cash, accrual]
        profitLoss:
          $ref: '#/components/schemas/YtdProfitLoss'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/IncomeTaxCategory'
        estimate:
          $ref: '#/components/schemas/IncomeTaxEstimate'
    TaxPeriod:
      type: object
      required: [start, end]
      properties:
        start:
          type: string
        end:
          type: string
    YtdProfitLoss:
      type: object
      required: [totalIncome, totalExpenses, profitLoss, taxableBusinessProfit, recordCount]
      properties:
        totalIncome:
          type: number
        totalExpenses:
          type: number
        profitLoss:
          type: number
        taxableBusinessProfit:
          type: number
        recordCount:
          type: number
    IncomeTaxCategory:
      type: object
      required: [category, recordType, amount, records]
      properties:
        category:
          type: string
        recordType:
          type: string
          enum: [income, expense]
        amount:
          type: number
        records:
          type: array
          items:
            $ref: '#/components/schemas/IncomeTaxRecord'
    IncomeTaxRecord:
      type: object
      required: [id, date, description, amount]
      properties:
        id:
          type: string
        date:
          type: string
        description:
          type: string
        amount:
          type: number
    IncomeTaxEstimate:
      type: object
      required: [status, taxYear, accountingBasis, businessProfitInput, otherTaxableIncome, totalIncome, personalAllowance, taxableIncome, estimatedIncomeTax, bands, assumptions, missingInputs]
      properties:
        status:
          type: string
          enum: [complete, incomplete]
        taxYear:
          type: string
        accountingBasis:
          type: string
          enum: [cash, accrual]
        businessProfitInput:
          type: number
        otherTaxableIncome:
          type: number
          nullable: true
        totalIncome:
          type: number
          nullable: true
        personalAllowance:
          type: number
          nullable: true
        taxableIncome:
          type: number
          nullable: true
        estimatedIncomeTax:
          type: number
          nullable: true
        bands:
          type: array
          items:
            $ref: '#/components/schemas/IncomeTaxBand'
        assumptions:
          type: array
          items:
            type: string
        missingInputs:
          type: array
          items:
            type: string
    IncomeTaxBand:
      type: object
      required: [label, rate, taxableAmount, tax]
      properties:
        label:
          type: string
        rate:
          type: number
        taxableAmount:
          type: number
        tax:
          type: number
...
    InboxSubOption:
      type: object
      required: [label]
      properties:
        label:
          type: string
        isSuggested:
          type: boolean
    InboxOption:
      type: object
      required: [label]
      properties:
        label:
          type: string
        isSuggested:
          type: boolean
        subOptions:
          type: array
          items:
            $ref: '#/components/schemas/InboxSubOption'
    InboxItem:
      type: object
      required: [id, profileId, date, description, status, aiReasoning, options, createdAt]
      properties:
        id:
          type: string
        profileId:
          type: string
        evidenceId:
          type: [string, 'null']
        date:
          type: string
        description:
          type: string
        amount:
          type: [number, 'null']
        aiReasoning:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/InboxOption'
        status:
          type: string
        resolution:
          type: [string, 'null']
        taxImpact:
          type: [number, 'null']
        createdAt:
          type: string
    InboxResolveInput:
      type: object
      required: [resolution]
      properties:
        resolution:
          type: string
          minLength: 1
...
    CopilotMessageInput:
      type: object
      required: [profileId, message]
      properties:
        profileId:
          type: string
        message:
          type: string
          minLength: 1
    CopilotMessageResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
        contextSummary:
          type: string
...